Removing react-based semantic ui, first includes of graphql
