+++ package.json
@@ -23,6 +23,7 @@
     "update-version": "node scripts/update-version.js"
   },
   "dependencies": {
+    "@apollo/client": "^3.3.14",
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.5.0",
@@ -30,13 +31,12 @@
     "crypto-js": "^4.0.0",
     "dayjs": "^1.9.7",
     "env-cmd": "^10.1.0",
+    "graphql": "^15.5.0",
     "node-sass": "^4.14.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^2.0.1",
     "web-vitals": "^0.2.4"
   },
   "devDependencies": {
