+++ src/components/core/home/home.scss
