+++ src/components/pets/pet-info/pet-info.jsx
