+++ src/components/shared/form-button/form-button.test.js
@@ -1,0 +1,12 @@
+import { render, screen } from '@testing-library/react'
+import FormButton from './form-button'
+
+test('Renders a FormButton using required props only', () => {
+  const props = {
+    type: 'button',
+    label: 'Test'
+  }
+  render(<FormButton config={props} />)
+  const element = screen.getByText(props.label)
+  expect(element).toBeInTheDocument()
+})
