+++ src/components/users/new-user-form/new-user-form.test.js
@@ -1,0 +1,8 @@
+import { render, screen } from '@testing-library/react'
+import NewUserForm from './new-user-form'
+
+test(`Renders a 'New User' Form without props`, () => {
+  render(<NewUserForm />)
+  const element = screen.getByText('Password')
+  expect(element).toBeInTheDocument()
+})
