+++ src/enums/buttons.enum.js
@@ -15,3 +15,20 @@ export const buttonColorEnums = [
   'grey',
   'black'
 ]
+
+export const columnWidth = {
+  1: 'one',
+  2: 'two',
+  3: 'three',
+  4: 'four',
+  5: 'five',
+  6: 'six',
+  7: 'seven',
+  8: 'eight',
+  9: 'nine',
+  10: 'ten',
+  11: 'eleven',
+  12: 'twelve',
+  13: 'thirteen',
+  14: 'fourteen'
+}
