--- src/tests/core/login-form.test.js
@@ -1,8 +1,0 @@
-import { render, screen } from '@testing-library/react'
-import LoginForm from '../../app/core/login-form/login-form'
-
-test(`Renders a 'Login' Form without props`, () => {
-  render(<LoginForm />)
-  const element = screen.getByText('Password')
-  expect(element).toBeInTheDocument()
-})
