--- src/tests/shared/grid-layout.test.js
@@ -1,8 +1,0 @@
-import { render, screen } from '@testing-library/react'
-import GridLayout from '../../app/shared/grid-layout/grid-layout'
-
-test('render a grid with required properties', () => {
-  render(<GridLayout header={{ title: 'Test' }} />)
-  const element = screen.getByText('Test')
-  expect(element).toBeInTheDocument()
-})
